bug fix from Tyler Nelson <pojojojo@gmail.com>
================== NOTE ============ 
I had to make a small change in SchemaBuilder.jar for the WS Commons  XmlSchema project. if was not working with a restriction tag in the amazon.com webservice  that I was testing out. Basically there was an exception with the QName because the name was not being set by the restriction tag logic correctly (I think?). I added code on line 344, in case you want to look over it.
 
Regards,
   Tyler
 
amazon webservice wsdl
http://webservices.amazon.com/AWSECommerceService/AWSECommerceService.wsdl
================== NOTE ============



git-svn-id: https://svn.apache.org/repos/asf/webservices/commons/trunk/XmlSchema@358648 13f79535-47bb-0310-9956-ffa450edef68